(feat): argon2i
I added support for argon2i using the same function I had created for
argon2id. I changed its signature to accept a variant. The variant
parameters takes an Algorithm, which is determined at the match branch
earlier in the file, which is responsible for calling the appropriate
KDF function based on the CLI. the argon2 function takes either an
Algorithm::2i variant or Algorithm::2id variant, and sets up argon2
accordingly.