(feat): verify command, output/input formats
I added a verify command to allow for verification of a hash given the
hash and salt. It takes these as inputs, uses the default values for
iterations, parallelism, etc if not provided, generates the hash given
these values and the provided salt, then compares the provided hash to
the generated hash.

I also added OutputFormat and InputFormat structs and associated CLI
args to allow for input and output in base64 or hex.